+++ src/main/java/cartelera/services/impl/CinemaServiceImpl.java
@@ -2,17 +2,19 @@ package cartelera.services.impl;
 
 import cartelera.entities.Cinema;
 import cartelera.entities.Room;
+import cartelera.repositories.AddressRepository;
 import cartelera.repositories.CinemaRepository;
-import cartelera.services.IAddressService;
+import cartelera.repositories.RoomRepository;
 import cartelera.services.ICinemaService;
-import cartelera.services.IRoomService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import static cartelera.utils.Utils.invalidPosNumber;
+
 @Slf4j
 @AllArgsConstructor
 @Service
@@ -20,8 +22,8 @@ import java.util.Optional;
 public class CinemaServiceImpl implements ICinemaService {
 
     private final CinemaRepository cinemaRepo;
-    private final IRoomService roomService;
-    private final IAddressService addressService;
+    private final RoomRepository roomRepo;
+    private final AddressRepository addressRepo;
 
     @Override
     public List<Cinema> findAll() {
@@ -30,33 +32,45 @@ public class CinemaServiceImpl implements ICinemaService {
 
     @Override
     public Optional<Cinema> findById(Long id) {
+        log.info("findById {}", id);
+        if (invalidPosNumber(id)) return Optional.empty();
         return cinemaRepo.findById(id);
     }
 
     @Override
+    public boolean existsById(Long id) {
+        log.info("existsById {}", id);
+        if (invalidPosNumber(id)) return false;
+        return cinemaRepo.existsById(id);
+    }
+
+    @Override
     public Cinema save(Cinema cinema) {
+        log.info("save {}", cinema);
         return cinemaRepo.save(cinema);
     }
 
     @Override
     public void deleteById(Long id) {
+        log.info("deleteById {}", id);
+        if (invalidPosNumber(id) && !cinemaRepo.existsById(id)) return;
+
+        Cinema cinema = findById(id).get();
 
         // borrar todas las rooms asociadas
-        List<Room> rooms = roomService.findAllByCinemaId(id);
-        List<Long> ids = new ArrayList<>();
-        for (Room room : rooms)
-            ids.add(room.getId());
-        roomService.deleteAllById(ids);
+        List<Room> rooms = roomRepo.findAllByCinema_Id(id);
+        if (!rooms.isEmpty()) for (Room room : rooms) roomRepo.deleteById(room.getId());
 
         // borrar address asociada
-        Optional<Cinema> cinemaOpt = findById(id);
-        if (cinemaOpt.isPresent()) {
-            Cinema cinema = cinemaOpt.get();
-            Long addressId = cinema.getAddress().getId();
-            cinema.setAddress(null);
-            addressService.deleteById(addressId);
-        }
+        Long addressId = cinema.getAddress().getId();
+        cinema.setAddress(null);
+        addressRepo.deleteById(addressId);
 
         cinemaRepo.deleteById(id);
+    }
+
+    @Override
+    public Optional<Cinema> findByIdWithRooms(Long id) {
+        return cinemaRepo.findByIdWithRooms(id);
     }
 }
